requirements.txt
# v2
alpaca-py
pandas
numpy
python-dotenv
pytz
flask
flask-cors
requests
gunicorn
